GUI functional and ready to be used.
run_segmentation.py was only working on selected movies (benchmarks) - fixed.

Added draft of class based segmentation script
Added older older version of run_segmentation.py for trubleshooting
