+++ .idea/.gitignore
@@ -6,3 +6,7 @@
 # Datasource local storage ignored files
 /dataSources/
 /dataSources.local.xml
+*.txt
+*.tif
+*.tiff
+*.nd2
