+++ run_segmentation.py
@@ -10,8 +10,8 @@
 #
 #
 # This file contains code to run cellpose segmentation on a tiff
-# movie, using our custom finetuned model that's adapted to 
-# Synechococcus sp. PCC 7002. This script is adapted to take in
+# movie, using our custom finetuned model that's adapted to [idk
+# cyanobacteria name spelling] 7002. This script is adapted to take in
 # a movie, convert the z-stack to single images as required by
 # cellpose, and then run segmentation. Output will then be coerced
 # from the cellpose output format (1 channel per cell) to a single
@@ -242,41 +242,80 @@ if denoise_p:
         device=device, model_type="denoise_cyto3"
     )
 
-# Select the file reader
-if file_type == "nd2":
-    reader = ND2Reader
-elif file_type == "tif":
-    reader = tifffile.imread
-
-
 def get_movie_frame(movie, frame_idx: int):
     """
     Given a movie and a frame, load the frame from the movie
     """
-    movie.bundle_axes = ["y", "x", "c"]
-    movie_frame = movie.get_frame(frame_idx)
+    if hasattr(movie, "bundle_axes"):
+        movie.bundle_axes = ["y", "x"]
+        movie_frame = movie.get_frame(frame_idx)
+    else:
+        movie_frame = movie[frame_idx]
     return np.array(movie_frame, dtype=np.uint16)
 
+# Run segmentation on all frames
+masks = []
+flows = []
+probs = []
 
 print(f"Reading input file {input_file}")
-with reader(input_file) as images:
+
+def nd2_seg(end_frame, size):
+    reader = ND2Reader
+    with reader(input_file) as images:
+        if end_frame == -1:
+            end_frame = len(images)
+        print(f"Running segmentation on {end_frame - start_frame} frames")
+        for i in tqdm(
+            range(start_frame, end_frame),
+            desc="Frames",
+            unit="frame",
+        ):
+            image = get_movie_frame(images, i)
+
+            #image = np.array(image)
+            if size is None:
+                size, size_style = size_model.eval(image, channels=chan)
+                print(f"Size estimated as {size} for frame {i}")
+                if size > 50:
+                    print(
+                        f"WARNING: Size estimated as {size}, this is unusually large"
+                    )
+            # Denoising
+            if denoise_p:
+                image = denoise_model.eval(image, channels=chan)
+            # Speed up with tile=False, uses more memory
+            mask, flow, _ = model.eval(
+                image,
+                diameter=size,
+                channels=chan,
+                #tile=True, # enabled by default in newer cellpose
+                niter=niter,
+                flow_threshold=flow_threshold,
+            )
+            flows.append(flow[0])
+            probs.append(flow[2])
+            masks.append(mask)
+
+def tif_seg(end_frame, size):
+    images = tifffile.imread(input_file)
+    print(images.shape)
+
     if end_frame == -1:
         end_frame = len(images)
-    # used_images = images[start_frame:end_frame]
-    print(f"Running segmentation on {end_frame - start_frame + 1} frames")
-    # Run segmentation on all frames
-    masks = []
-    flows = []
-    probs = []
+    print(f"Running segmentation on {end_frame - start_frame} frames")
+
     # Run on single core or GPU if available
     # TODO Fix to use minibatches
     for i in tqdm(
-        range(start_frame, end_frame + 1),
+        range(start_frame, end_frame),
         desc="Frames",
         unit="frame",
     ):
         image = get_movie_frame(images, i)
-        # image = np.array(image)
+        image = image[0, :, :]
+        print(image.shape)
+        #image = np.array(image)
         if size is None:
             size, size_style = size_model.eval(image, channels=chan)
             print(f"Size estimated as {size} for frame {i}")
@@ -292,7 +331,7 @@ with reader(input_file) as images:
             image,
             diameter=size,
             channels=chan,
-            tile=True,
+            #tile=True, # enabled by default in newer cellpose
             niter=niter,
             flow_threshold=flow_threshold,
         )
@@ -300,25 +339,31 @@ with reader(input_file) as images:
         probs.append(flow[2])
         masks.append(mask)
 
-    # Stack the masks
-    masks = np.stack(masks)
-    # Save the mask
-    print(f"Saving to {output_file}")
-    # Coerce to single channel
-    # masks = np.clip(masks, 0, 1)
-    # Manually save the masks
-    tifffile.imwrite(output_file, masks)
-    if debug:
-        print("Saving debug files")
-        flows = np.stack(flows)
-        # Save the flows
-        flow_file = output_file + ".flow.tif"
-        print(f"Saving flows to {flow_file}")
-        tifffile.imwrite(flow_file, flows, bigtiff=True)
-        # Save the probs
-        prob_file = output_file + ".prob.tif"
-        print(f"Saving probs to {prob_file}")
-        tifffile.imwrite(prob_file, probs, bigtiff=True)
+
+if file_type == "nd2":
+    nd2_seg(end_frame, size)
+elif file_type == "tif":
+    tif_seg(end_frame, size)
+
+# Stack the masks
+masks = np.stack(masks)
+# Save the mask
+print(f"Saving to {output_file}")
+# Coerce to single channel
+
+# Manually save the masks
+tifffile.imwrite(output_file, masks)
+if debug:
+    print("Saving debug files")
+    flows = np.stack(flows)
+    # Save the flows
+    flow_file = output_file + ".flow.tif"
+    print(f"Saving flows to {flow_file}")
+    tifffile.imwrite(flow_file, flows, bigtiff=True)
+    # Save the probs
+    prob_file = output_file + ".prob.tif"
+    print(f"Saving probs to {prob_file}")
+    tifffile.imwrite(prob_file, probs, bigtiff=True)
 
 
 #
